# Enable reading in reverse order optimization for FINAL queries