--- a/src/Processors/Merges/ReplacingSortedTransform.h
+++ b/src/Processors/Merges/ReplacingSortedTransform.h
@@ -20,7 +20,8 @@ public:
         WriteBuffer * out_row_sources_buf_ = nullptr,
         bool use_average_block_sizes = false,
         bool cleanup = false,
-        bool enable_vertical_final = false)
+        bool enable_vertical_final = false,
+        int read_direction = 1)
         : IMergingTransform(
             num_inputs, header, header, /*have_all_inputs_=*/ true, /*limit_hint_=*/ 0, /*always_read_till_end_=*/ false,
             header,
@@ -33,7 +34,8 @@ public:
             out_row_sources_buf_,
             use_average_block_sizes,
             cleanup,
-            enable_vertical_final)
+            enable_vertical_final,
+            read_direction)
     {
     }