--- a/tests/queries/0_stateless/02564_read_in_order_final_desc.sql
+++ b/tests/queries/0_stateless/02564_read_in_order_final_desc.sql
@@ -27,7 +27,7 @@ SELECT if(explain like '%ReadType: InOrder%', 'Ok', 'Error: ' || explain) FROM (
 ) WHERE explain like '%ReadType%';
 
 
-SELECT if(explain like '%ReadType: Default%', 'Ok', 'Error: ' || explain) FROM (
+SELECT if(explain like '%ReadType: InReverseOrder%', 'Ok', 'Error: ' || explain) FROM (
     EXPLAIN PLAN actions = 1
     SELECT timestamp, value
     FROM mytable FINAL