… tests/queries/0_stateless/02947_read_in_reverse_order_final.reference