Enable reading in reverse order when sorting against the sorting key in FINAL queries. Addresses slowness explained in https://github.com/ClickHouse/ClickHouse/issues/58035. Closes https://github.com/ClickHouse/ClickHouse/issues/58035.

The reason for previously disabling it seems to have been some concerns about incorrect behavior of merging algorithms (see deleted comment). But no specific incorrect behavior is mentioned, and all the tests I wrote seem to work fine. Maybe CI will find something (I don't have a full setup myself that can run all tests).

As expected, performance is much improved for the queries mentioned in https://github.com/ClickHouse/ClickHouse/issues/58035:
```sql
arrakis :) SELECT x FROM numbers FINAL ORDER BY x DESC LIMIT 1 SETTINGS optimize_read_in_order = 0

┌─────────x─┐
│ 499999999 │
└───────────┘

1 row in set. Elapsed: 8.583 sec. Processed 500.00 million rows, 2.00 GB (58.25 million rows/s., 233.01 MB/s.)
Peak memory usage: 8.13 MiB.

arrakis :) SELECT x FROM numbers FINAL ORDER BY x DESC LIMIT 1 SETTINGS optimize_read_in_order = 1

┌─────────x─┐
│ 499999999 │
└───────────┘

1 row in set. Elapsed: 1.183 sec. Processed 793.56 thousand rows, 3.17 MB (670.57 thousand rows/s., 2.68 MB/s.)
Peak memory usage: 1.88 GiB.
```

### Changelog category (leave one):
- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Enable reading in reverse order optimization for FINAL queries.