@cwurm 2023-12-30
Reading in reverse order for FINAL was disabled in https://github.com/ClickHouse/ClickHouse/pull/46009 following bug report https://github.com/ClickHouse/ClickHouse/issues/45815. The example provided is working fine with this implementation. FYI @vdimir