@canhld94 2024-01-02
As in https://github.com/ClickHouse/ClickHouse/issues/58035#issuecomment-1867235883, I can foresee two cases that read-in-reverse-order may lead to wrong FINAL behaviour:

1. ReplacingMergeTree
For example, we read from a level-0 part with two rows 1 and 10 that have duplicated keys. If we use read-in-order, row 10 will be kept (expected); if we use read-in-reverse-order, row 1 will be kept
2. AggregatingMergeTree
For example, same situation as above, and aggregate function is `anyFirst`, `anyLast`, `first_value`, `last_value`, and depending on the data, `argMin`, `argMax` may also have issue.

I cannot figure out if there is any issue with `CollapsingMergeTree` or `VersionedCollapsingMergeTree` yet (we usually don't use FINAL for such engines).