@cwurm 2024-01-02
> As in [#58035 (comment)](https://github.com/ClickHouse/ClickHouse/issues/58035#issuecomment-1867235883), I can foresee two cases that read-in-reverse-order may lead to wrong FINAL behaviour:
> 
> 1. ReplacingMergeTree
>    For example, we read from a level-0 part with two rows 1 and 10 that have duplicated keys. If we use read-in-order, row 10 will be kept (expected); if we use read-in-reverse-order, row 1 will be kept

Indeed, you're right. I need to think about what to do about this case and others. For now, I've added a test for it that will fail with the current implementation.

> Simple way to bypass these issues is to only use read-in-reverse-order if there is no level-0 part in source parts.

Yeah, that could be the easy way to go.