@UnamedRus 2024-01-02
> and depending on the data, argMin, argMax may also have issue.

They aren't stable if you have the same number for value.
So, it's kinda OK